+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 
 * nostr: add optional relay URL arg to `Tag::coordinate` ([Yuki Kishimoto])
 * nostr: update `TagStandard::Label` and `EventBuilder::label` ([Yuki Kishimoto])
+* nostr: update `custom` field type in `Metadata` struct ([Yuki Kishimoto])
 * pool: remove `Error::Failed` variant ([Yuki Kishimoto])
 * pool: returns `Output` instead of an error if the message/event sending fails for all relays ([Yuki Kishimoto])
 * pool: add `reason` field to `AdmitStatus::Rejected` variant ([Yuki Kishimoto])
@@ -54,6 +55,7 @@
 * nostr: add `push`, `pop`, `insert` and `extend` methods to the `Tag` struct ([Yuki Kishimoto])
 * nostr: add `nip47::Notification` ([daywalker90])
 * nostr: add `MachineReadablePrefix::as_str` method ([Yuki Kishimoto])
+* nostr: derive `Hash` for `Metadata` ([Yuki Kishimoto])
 * pool: add `Relay::ban` method ([Yuki Kishimoto])
 * pool: add `AdmitPolicy::admit_connection` method ([Yuki Kishimoto])